Rename soulsmods for consistency
