+++ rust-toolchain.toml
@@ -1,3 +1,2 @@
 [toolchain]
-channel = "nightly-2022-06-22" 
 targets = [ "i686-pc-windows-msvc", "x86_64-pc-windows-msvc" ]
