+++ src/SoulMemory/SoulMemory.csproj
@@ -44,13 +44,13 @@
     <None Remove="resources\**" />
   </ItemGroup>
   <ItemGroup>
-    <None Remove="soulsmods\x64\soulsmods.dll" />
+    <None Remove="soulmods\x64\soulmods.dll" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="soulsplitter.ico" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="soulsmods\x64\soulsmods.dll" />
+    <EmbeddedResource Include="soulmods\x64\soulmods.dll" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
