+++ src/soulmods/src/console.rs
@@ -1,0 +1,29 @@
+// This file is part of the soulmemory-rs distribution (https://github.com/FrankvdStam/soulmemory-rs).
+// Copyright (c) 2022 Frank van der Stam.
+// https://github.com/FrankvdStam/soulmemory-rs/blob/main/LICENSE
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, version 3.
+//
+// This program is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+use windows::Win32::System::Console::{AllocConsole, FreeConsole};
+
+
+pub fn init_console()
+{
+    unsafe{ AllocConsole().unwrap() };
+}
+
+#[allow(dead_code)]
+pub fn free_console()
+{
+    unsafe{ FreeConsole().unwrap() };
+}
