--- src/soulsmods/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "soulsmods"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-detour = "0.8.1"
-ilhook = "2.0.0"
-
-mem-rs = "0.1.3"
-#mem-rs = { path="C:/projects/mem-rs" }
-lazy_static = "1.4.0"
-
-log = "0.4.17"
-log4rs = {version = "1.1.1", features = ["all_components" ] }
-
-[dependencies.windows]
-version = "0.51.1"
-features = [
-    "Win32_Foundation",
-    "Win32_System_Memory",
-    "Win32_System_Diagnostics_Debug",
-    "Win32_System_ProcessStatus",
-    "Win32_System_Threading",
-    "Win32_System_Console",
-    "Win32_System_SystemServices",
-    "Win32_System_LibraryLoader",
-    "Win32_Security",
-    "Win32_UI_Input_XboxController",
-]
